+++ JML/JML.ApiModels/KnowledgeTestModel.cs
@@ -1,6 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Text;
 
 namespace JML.ApiModels
 {
