+++ JML/JML.BusinessLogic.Core/Contracts/Emails/IEmailService.cs
@@ -1,6 +1,5 @@
 ﻿using System.Threading.Tasks;
 using JML.ApiModels;
-using JML.Domain;
 
 namespace JML.BusinessLogic.Core.Contracts.Emails
 {
