+++ JML/JML.BusinessLogic.Core/Contracts/Groups/IGroupService.cs
@@ -10,5 +10,6 @@ namespace JML.BusinessLogic.Core.Contracts.Groups
         Task<GroupModel> GetSimple(Guid id);
         Task<List<GroupModel>> GetAllSimple();
         Task<GroupModel> Create(string name);
+        Task Update(UpdateGroupModel model);
     }
 }
