+++ JML/JML.BusinessLogic.Core/Contracts/KnowledgeTests/IKnowledgeTestResultService.cs
@@ -1,0 +1,11 @@
+﻿using JML.ApiModels;
+using JML.Domain;
+
+namespace JML.BusinessLogic.Core.Contracts.KnowledgeTests
+{
+    public interface IKnowledgeTestResultService
+    {
+        bool IsProvidedCorrectAnswer(KnowledgeTestQuestion question, KnowledgeQuestionModel model);
+        KnowledgeTestResultModel GetResult(KnowledgeTest test);
+    }
+}
