+++ JML/JML.BusinessLogic.Core/Contracts/Lectures/ILectureService.cs
@@ -7,6 +7,7 @@ namespace JML.BusinessLogic.Core.Contracts.Lectures
 {
     public interface ILectureService
     {
+        Task<List<SectionGroupModel>> GetAll();
         Task<LectureModel> GetAsync(string url);
         Task<List<LectureModel>> GetAsync();
         Task<LectureModel> CreateAsync(LectureModel model);
