+++ JML/JML.BusinessLogic.Core/Contracts/Lectures/ILiteratureService.cs
@@ -1,0 +1,11 @@
+﻿using JML.ApiModels;
+using System.Threading.Tasks;
+
+namespace JML.BusinessLogic.Core.Contracts.Lectures
+{
+    public interface ILiteratureService
+    {
+        Task<LiteratureModel> Get();
+        Task Update(string content);
+    }
+}
