+++ JML/JML.Presentation.WebClient/Controllers/GroupsController.cs
@@ -1,13 +1,9 @@
 ﻿using JML.ApiModels;
 using JML.BusinessLogic.Core.Contracts.Groups;
-using JML.DataAccess.Core.Contracts;
-using JML.Domain;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace JML.Presentation.WebClient.Controllers
@@ -18,19 +14,10 @@ namespace JML.Presentation.WebClient.Controllers
     public class GroupsController : ControllerBase
     {
         private readonly IGroupService groupService;
-        private readonly IAppEntityRepository<StudyGroup> studyGroupsRepository;
-        private readonly IAppEntityRepository<User> usersRepository;
-        private readonly IDataContext dataContext;
 
-        public GroupsController(IGroupService groupService,
-            IAppEntityRepository<StudyGroup> studyGroupsRepository,
-            IAppEntityRepository<User> usersRepository,
-            IDataContext dataContext)
+        public GroupsController(IGroupService groupService)
         {
             this.groupService = groupService;
-            this.studyGroupsRepository = studyGroupsRepository;
-            this.usersRepository = usersRepository;
-            this.dataContext = dataContext;
         }
 
         [HttpGet]
@@ -51,31 +38,15 @@ namespace JML.Presentation.WebClient.Controllers
         [HttpPost]
         public async Task<ActionResult<GroupModel>> Create(GroupModel model)
         {
-            return Ok(await groupService.Create(model.Name));
+            var group = await groupService.Create(model.Name);
+            return Ok(group);
         }
 
         [HttpPost]
         [Route("update")]
         public async Task<ActionResult> Update(UpdateGroupModel model)
         {
-            var group = await studyGroupsRepository.GetQuery().FirstAsync(x => x.Id == model.Id);
-
-            if (group != null)
-            {
-                group.Name = model.Name;
-                
-                foreach (var user in await usersRepository.GetQuery().ToListAsync())
-                {
-                    user.GroupId = null;
-                    if (model.Users != null && model.Users.Contains(user.Id))
-                    {
-                        user.GroupId = group.Id;
-                    }
-                }
-
-                await dataContext.SaveChangesAsync();
-            }
-
+            await groupService.Update(model);
             return Ok();
         }
     }
