+++ JML/JML.Presentation.WebClient/Controllers/LiteratureController.cs
@@ -1,14 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+﻿using System.Threading.Tasks;
 using JML.ApiModels;
-using JML.DataAccess.Core.Contracts;
-using JML.Domain;
+using JML.BusinessLogic.Core.Contracts.Lectures;
 using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
 
 namespace JML.Presentation.WebClient.Controllers
 {
@@ -17,43 +11,24 @@ namespace JML.Presentation.WebClient.Controllers
     [Route("api/[controller]")]
     public class LiteratureController : ControllerBase
     {
-        private readonly IAppEntityRepository<Literature> literatureRepository;
-        private readonly IDataContext dataContext;
+        private readonly ILiteratureService literatureService;
 
-        public LiteratureController(IAppEntityRepository<Literature> literatureRepository,
-            IDataContext dataContext)
+        public LiteratureController(ILiteratureService literatureService)
         {
-            this.literatureRepository = literatureRepository;
-            this.dataContext = dataContext;
+            this.literatureService = literatureService;
         }
 
         [HttpGet]
         public async Task<ActionResult> Get()
         {
-            var literature = await literatureRepository.GetQuery().FirstOrDefaultAsync();
-            var model = new LiteratureModel
-            {
-                Content = literature?.Content ?? String.Empty,
-                ModifiedAt = literature?.ModifiedAt ?? DateTime.Now,
-            };
-
-            return Ok(model);
+            var literature = await literatureService.Get();
+            return Ok(literature);
         }
 
         [HttpPost]
         public async Task<ActionResult> Post(LiteratureModel model)
         {
-            var literature = await literatureRepository.GetQuery().FirstOrDefaultAsync();
-
-            if (literature == null)
-            {
-                literature = new Literature();
-                literatureRepository.Add(literature);
-            }
-
-            literature.Content = model.Content;
-            await dataContext.SaveChangesAsync();
-
+            await literatureService.Update(model.Content);
             return Ok();
         }
     }
